# stringutil: improve DoMatch performance