<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

`DoMatch` has bad performance on special case.

related to https://github.com/tikv/tikv/pull/5866


### What is changed and how it works?

Optimize the DoMatch algorithm by https://research.swtch.com/glob

Backtrace only back to last `patAny`

All characters will be traversed at most 2 times.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

I add a new special case benchmark test.

Before optimization:

```
PASS: string_util_test.go:128: testStringUtilSuite.TestIsExactMatch	0.000s
PASS: string_util_test.go:80: testStringUtilSuite.TestPatternMatch	0.000s
PASS: string_util_test.go:33: testStringUtilSuite.TestUnquote	0.000s
OK: 3 passed
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/util/stringutil
BenchmarkMatchSpecial-8   	       3	 426341220 ns/op
PASS
ok  	github.com/pingcap/tidb/util/stringutil	2.604s

```

After optimization:

```
PASS: string_util_test.go:128: testStringUtilSuite.TestIsExactMatch	0.000s
PASS: string_util_test.go:80: testStringUtilSuite.TestPatternMatch	0.000s
PASS: string_util_test.go:33: testStringUtilSuite.TestUnquote	0.000s
OK: 3 passed
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/util/stringutil
BenchmarkMatchSpecial-8   	 8777878	       135 ns/op
PASS
ok  	github.com/pingcap/tidb/util/stringutil	1.336s

```

 - Unit test

Code changes

Side effects

Related changes

Release note

 - Optimize DoMatch performance