@zz-jason 2019-11-13
@TennyZhuang It's better to add some benchmarks for the performance improvement. Could you add one in a test file to tell the code reviewers about how much performance does this PR improved?