@TennyZhuang 2019-11-13
@zz-jason I have add a benchmark, and add my result in description.