@TennyZhuang 2019-11-22
@wshwsh12 I have updated my benchmark result, PTAL