@TennyZhuang 2019-11-22
And the corresponding optimization is merged in https://github.com/tikv/tikv/pull/5866